design-mvp/requirements.txt
fastapi>=0.109.0
python-dotenv>=1.0.0
openai>=1.12.0
Pillow>=10.2.0
uvicorn>=0.27.0
python-multipart>=0.0.7
langchain-openai>=0.0.5